moves to session folder
