+++ services/web/server/src/simcore_service_webserver/login/handlers_2fa.py
@@ -11,7 +11,7 @@ from servicelib.error_codes import create_error_code
 from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
 
 from ..products.plugin import Product, get_current_product
-from ..session_access import session_access_required
+from ..session.session_access import session_access_required
 from ._2fa import (
     create_2fa_code,
     get_2fa_code,
