+++ services/web/server/src/simcore_service_webserver/login/handlers_confirmation.py
@@ -15,7 +15,7 @@ from yarl import URL
 
 from ..products.plugin import Product, get_current_product
 from ..security.api import encrypt_password
-from ..session_access import session_access_required
+from ..session.session_access import session_access_required
 from ..utils import MINUTE
 from ..utils_aiohttp import create_redirect_response
 from ..utils_rate_limiting import global_rate_limit_route
