+++ services/web/server/src/simcore_service_webserver/session/settings.py
@@ -7,7 +7,7 @@ from pydantic.types import SecretStr
 from settings_library.base import BaseCustomSettings
 from settings_library.utils_session import MixinSessionSettings
 
-from ._constants import APP_SETTINGS_KEY
+from .._constants import APP_SETTINGS_KEY
 
 _MINUTES: Final[int] = 60  # secs
