+++ services/web/server/tests/unit/isolated/test_activity.py
@@ -17,7 +17,7 @@ from simcore_service_webserver.activity.plugin import setup_activity
 from simcore_service_webserver.application_settings import setup_settings
 from simcore_service_webserver.rest import setup_rest
 from simcore_service_webserver.security.plugin import setup_security
-from simcore_service_webserver.session import setup_session
+from simcore_service_webserver.session.plugin import setup_session
 
 
 @pytest.fixture
